backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication & Storage
supabase==2.9.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# AI Services
google-generativeai==0.3.2

# Background Tasks
celery[redis]==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0
